+++ include/openpfc/openpfc.hpp
@@ -38,8 +38,6 @@
 #include <openpfc/kernel/fft/fft_fftw.hpp>
 #include <openpfc/kernel/mpi/mpi.hpp>
 #include <openpfc/kernel/simulation/binary_reader.hpp>
-#include <openpfc/kernel/simulation/boundary_conditions/fixed_bc.hpp>
-#include <openpfc/kernel/simulation/boundary_conditions/moving_bc.hpp>
 #include <openpfc/kernel/simulation/field_modifier.hpp>
 #include <openpfc/kernel/simulation/initial_conditions/constant.hpp>
 #include <openpfc/kernel/simulation/initial_conditions/file_reader.hpp>
@@ -51,7 +49,7 @@
 #include <openpfc/kernel/simulation/results_writer.hpp>
 #include <openpfc/kernel/simulation/simulation_context.hpp>
 #include <openpfc/kernel/simulation/simulator.hpp>
-#include <openpfc/kernel/simulation/time.hpp>
 #include <openpfc/kernel/simulation/steppers/butcher_tableau.hpp>
+#include <openpfc/kernel/simulation/time.hpp>
 
 #endif // PFC_OPENPFC_HPP